UI: change name to open educational badges (#9)
* Replace MINT-o-Badges

Replaced all occurrences of the text MINT-o-Badges and similar symbols
with Open Educational Badges (or similar where appropriate).
This does not affect links / mail-addresses. There I just added a TODO.
Also note that the logos are still left unchanged, thus partially still
displaying "my Badges".

* Replace logos

I replaced the text in the logos with "Open Educational Badges" in
InkScape. It's not beautiful, but for now it should do.

---------

Authored-by: Timber Borner <timber@opensenselab.org>