[WIP] UI create curtian newsletter (#34)
* Basic hacky curtain

* Move curtain to allow impressum

Moved the curtain to `app.component.html`, to still show the footer,
containing the impressum etc. Also don't show the curtain when the url
contains "impressum".

* Disable curtain with query parameter

The curtain can be disabled with the query parameter
`?curtainEnabled=false`. The disabled state is then stored in the local
storage, until either the local storage is cleared, or the url is called
with the query parameter `?curtainEnabled=true`.

* Remove unneeded curtain class

* Change color and font of curtain

* Add newsletter, not nice yet

* add curtain background

* add media query

* add line height for mobile

---------

Co-authored-by: Timber <timber@opensenselab.org>