Feat/oauth meinbildungsraum (#406)
* WIP

* Validate token functionality in the frontend (not yet automatised)

* Validate token automatically and fix some errors

* Also logout from mbr on oeb logout

If the user logged in via mbr and they logout via oeb, this (together
with the changes in badgr-server) will redirect them to a logout page
of mbr, where they can confirm to also logout from mbr and then to go
back to oeb.

* Change logout redirect endpoint

* Remove duplicates

* Prevent accidental navigation to token validation

If the user now access the login page with the ?validateToken query
parameter and the backend returns a 401 error, the user is redirected to
the same page without the query parameter and now error is shown. This
is because it can happen quite easily that the user navigates to the
address containing the query parameter accidentally.
In all other cases, e.g. when there were too many retries, the error
message is still shown (and the user is not redirected).

* Loud token renewal

* Use iframe for silent token renewal

This utilizes the prompt=none query parameter for the backend which is
passed via mozilla-django-oidc to prevent illegal frame renderings. Now
the only issue that remains to be solved is the storing of the cookies,
which get interpreted as third party cookies...

* Revert "Use iframe for silent token renewal"

This reverts commit b576924511075bd9591f6ca857abc7eb18f707f7.

* Revert "Loud token renewal"

This reverts commit 1684122bf29cc9b9cd2147c640f6e8e7c66478a9.

* Functional refresh token implementation

* fix: merge error

Re-remove old mbr button

* fix: re-add function to mbr button

---------

Co-authored-by: Timber <timber@opensenselab.org>